+++ NEWS.md
@@ -1,6 +1,7 @@
 # bspm 0.3.4.1
 
 - Fix noise when `busctl` is installed but there's no system bus (docker) (#12).
+- Workaround issue with file permissions under `/tmp` (#13).
 
 # bspm 0.3.4